pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name="edbo_data"
version="0.0.1"
authors = [
  { name="Mats Gustafsson", email="matsgus71@gmail.com" },
]
description = "My Python template"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python_support@git+https://git@github.com/eragnms/python_support.git@main",
    "smhi_pkg",
    "aiohttp",
]
[project.scripts]
edbo-data = "edbo_data.edbo_data:main"
